Factorize code & upgrade MegaLinter to v5
