+++ config/sfdx-hardis.mega-linter-config.yml
@@ -1,5 +1,5 @@
 # Configuration file for Mega-Linter
-# See all available variables at https://nvuillam.github.io/mega-linter/configuration/ and in linters documentation
+# See all available variables at https://megalinter.github.io/configuration/ and in linters documentation
 
 # To use this config from your local .mega-linter.yml file, add the following lines:
 # EXTENDS:
