+++ defaults/ci/.gitlab-ci.yml
@@ -63,11 +63,11 @@ check_quality:
   except:
     variables:
       - $SCRATCH_ORG_POOL == "true"
-  image: nvuillam/mega-linter-salesforce:latest
+  image: megalinter/megalinter-salesforce:beta
   script: ["true"]
   variables:
     # All available variables are described in documentation
-    # https://nvuillam.github.io/mega-linter/configuration/
+    # https://megalinter.github.io/configuration/
     DEFAULT_WORKSPACE: $CI_PROJECT_DIR
     DEFAULT_BRANCH: master
     # ADD YOUR CUSTOM ENV VARIABLES HERE TO OVERRIDE VALUES OF .mega-linter.yml AT THE ROOT OF YOUR REPOSITORY
